fix(snapshot): 上线前 review 收敛：评论作者标记、多选返回键、列表回顶、导入双写
review 标的是 #1046 离线快照 v1 合并后的最终状态（含后续 i18n 提交）。

- V3 快照评论页「作者」标记恒不亮：commentsRenderer 把 illustAuthorId(Int) 塞进
  intent，而 CommentsFragment 的 args 走 getLong 读，拿到的是默认值 0，
  CommentFeedItem.isAuthor 因此恒 false。经典详情页那条同名入口传的是 Long，
  只有 V3 这一侧漏了，显式 toLong()。
- 批量删除的 IO 循环里还在 requireContext()：转屏 detach 会抛 ISE，而它是裸
  launch，逃逸出去就是崩进程、还删了一半。同文件的 import / export 已经为这个坑
  改过并留了注释，delete 补上；列表页三处裸 launch 一并兜住（openSnapshot 调的
  loadViewerData 对「卡片指向的快照已被删」是会抛的）。
- 多选态按系统返回键直接关页：本页是仿 FragmentHistoryTabs 做的，那边有
  OnBackPressedCallback 优先退选择态，这里漏了。callback 只在选择态 enabled。
- 列表 onResume 的例行刷新也走「清空重建 + 回顶」：看完一个快照返回就闪一下并跳回
  顶部。回顶是导入后为保证新卡排序才要的，收敛成 reload(resetScroll) 显式参数。
- 导入把整份快照写两遍：解压到 cacheDir 之后再 copyRecursively 到 filesDir，两者
  同挂载点，改成先试 renameTo、失败再回落拷贝。这正是「导入不再先落一份完整 ZIP」
  那次要省掉的同一笔开销。
- artworkPageRenderer 在快照分支重复覆盖 itemView 点击：adapter 已按 snapshotId
  路由到「快照大图」且带了 EXTRA_ENTER_BOUNDS，覆盖后大图页进场退化成居中淡入。
- snapshotExtension() 取值顺序写反（先取 `?` 之后的东西 = 查询串本身），带 query
  的 URL 一律存成 .img。
- 列表卡片每次 bind 现构造 SimpleDateFormat，提成常量。
- 补 SnapshotPathSafetyTest：快照的路径安全边界（requireSnapshotId / safeResolve /
  unzip 的 `..` 过滤）全是纯 JVM 逻辑，且输入来自用户导入的外部 ZIP，值得钉住。

+++ app/src/main/java/ceui/pixiv/ui/detail/ArtworkSectionRenderers.kt
@@ -677,8 +677,11 @@ internal fun ArtworkV3Fragment.commentsRenderer() =
                     return
                 }
                 intent.putExtra(TemplateActivity.EXTRA_FRAGMENT, "快照评论")
-                intent.putExtra("objectId", item.illustId)
-                intent.putExtra("objectArthurId", item.illustAuthorId)
+                // 必须显式转 Long：CommentsFragment 的 args 是 getLong 读的，
+                // 塞 Int 进去 getLongExtra 只会拿到默认值 0 —— objectArthurId 一旦是 0，
+                // CommentFeedItem.isAuthor 恒 false，作者本人的评论就不再有「作者」标记。
+                intent.putExtra("objectId", item.illustId.toLong())
+                intent.putExtra("objectArthurId", item.illustAuthorId.toLong())
                 intent.putExtra("objectType", ceui.loxia.ObjectType.ILLUST)
                 intent.putExtra(
                     ceui.pixiv.snapshot.SnapshotManagerFragment.ARG_SNAPSHOT_ID,
